--- lua/plugins.old/telescope_nvim.lua
@@ -1,12 +1,0 @@
-return {
-	name = 'telescope.nvim',
-	dir = vim.g.nixplugins['telescope.nvim'],
-	event = 'VeryLazy',
-
-	dependencies = {
-		'plenary.nvim',
-	},
-
-	config = true,
-}
-
