--- lua/plugins.old/vsnip.lua
@@ -1,6 +1,0 @@
-return {
-	name = 'vim-vsnip',
-	dir = vim.g.nixplugins['vim-vsnip'],
-	event = 'VeryLazy',
-	main = 'vsnip',
-}
